# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Hotel Reservation Back-Office API
  version: 1.0.0
  description: API for managing hotel reservations, users, and payments.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /users:
    get:
      summary: Get all users
      responses:
        '200':
          description: Returns a list of users
    put:
      summary: Update user details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: User updated successfully
  /users/{id}:
    get:
      summary: Get user details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns user details
...